requirements.txt
Django
django-autoslug
Pillow
django-environ
django-hitcount
psycopg2
locust


